Update HSeriesPost.py
Added the ability for the user to choose how far in advance tools pre-heat. They enter an integer representing how many lines before the tool-change the tool will pre-heat. Entering 0 will disable pre-heating. If the user enters a number that is too high, the code will put the pre-heat as far ahead as possible.